tensorNet abstract class with GIE
